apply user auth to all clis
