+++ src/py/flwr/cli/run/run.py
@@ -17,7 +17,6 @@
 import io
 import json
 import subprocess
-from logging import DEBUG
 from pathlib import Path
 from typing import Annotated, Any, Optional, Union
 
@@ -25,23 +24,19 @@ import typer
 from rich.console import Console
 
 from flwr.cli.build import build
-from flwr.cli.cli_interceptor import CliInterceptor
 from flwr.cli.config_utils import (
     get_fab_metadata,
     load_and_validate,
-    validate_certificate_in_federation_config,
     validate_federation_in_project_config,
     validate_project_config,
 )
-from flwr.common.auth_plugin import CliAuthPlugin
 from flwr.common.config import (
     flatten_dict,
     parse_config_args,
     user_config_to_configsrecord,
 )
 from flwr.common.constant import CliOutputFormat
-from flwr.common.grpc import GRPC_MAX_MESSAGE_LENGTH, create_channel
-from flwr.common.logger import log, redirect_output, remove_emojis, restore_output
+from flwr.common.logger import redirect_output, remove_emojis, restore_output
 from flwr.common.serde import (
     configs_record_to_proto,
     fab_to_proto,
@@ -52,14 +47,9 @@ from flwr.proto.exec_pb2 import StartRunRequest  # pylint: disable=E0611
 from flwr.proto.exec_pb2_grpc import ExecStub
 
 from ..log import start_stream
-from ..utils import try_obtain_cli_auth_plugin
+from ..utils import init_channel, try_obtain_cli_auth_plugin
 
 CONN_REFRESH_PERIOD = 60  # Connection refresh period for log streaming (seconds)
-
-
-def on_channel_state_change(channel_connectivity: str) -> None:
-    """Log channel connectivity."""
-    log(DEBUG, channel_connectivity)
 
 
 # pylint: disable-next=too-many-locals
@@ -117,14 +107,13 @@ def run(
         )
 
         if "address" in federation_config:
-            auth_plugin = try_obtain_cli_auth_plugin(app, federation, federation_config)
             _run_with_exec_api(
                 app,
+                federation,
                 federation_config,
                 config_overrides,
                 stream,
                 output_format,
-                auth_plugin,
             )
         else:
             _run_without_exec_api(app, federation_config, config_overrides, federation)
@@ -148,23 +137,14 @@ def run(
 # pylint: disable-next=R0913, R0914, R0917
 def _run_with_exec_api(
     app: Path,
+    federation: str,
     federation_config: dict[str, Any],
     config_overrides: Optional[list[str]],
     stream: bool,
     output_format: str,
-    auth_plugin: Optional[CliAuthPlugin] = None,
 ) -> None:
-    insecure, root_certificates_bytes = validate_certificate_in_federation_config(
-        app, federation_config
-    )
-    channel = create_channel(
-        server_address=federation_config["address"],
-        insecure=insecure,
-        root_certificates=root_certificates_bytes,
-        max_message_length=GRPC_MAX_MESSAGE_LENGTH,
-        interceptors=(CliInterceptor(auth_plugin) if auth_plugin is not None else None),
-    )
-    channel.subscribe(on_channel_state_change)
+    auth_plugin = try_obtain_cli_auth_plugin(app, federation, federation_config)
+    channel = init_channel(app, federation_config, auth_plugin)
     stub = ExecStub(channel)
 
     fab_path, fab_hash = build(app)
