+++ src/py/flwr/cli/stop.py
@@ -15,24 +15,21 @@
 """Flower command line interface `stop` command."""
 
 
-from logging import DEBUG
 from pathlib import Path
-from typing import Annotated, Any, Optional
+from typing import Annotated, Optional
 
-import grpc
 import typer
 
 from flwr.cli.config_utils import (
     load_and_validate,
-    validate_certificate_in_federation_config,
     validate_federation_in_project_config,
     validate_project_config,
 )
 from flwr.common.constant import FAB_CONFIG_FILE
-from flwr.common.grpc import GRPC_MAX_MESSAGE_LENGTH, create_channel
-from flwr.common.logger import log
 from flwr.proto.exec_pb2 import StopRunRequest, StopRunResponse  # pylint: disable=E0611
 from flwr.proto.exec_pb2_grpc import ExecStub
+
+from .utils import init_channel, try_obtain_cli_auth_plugin
 
 
 def stop(
@@ -70,7 +67,8 @@ def stop(
         raise typer.Exit(code=1)
 
     try:
-        channel = _init_channel(app, federation_config)
+        auth_plugin = try_obtain_cli_auth_plugin(app, federation, federation_config)
+        channel = init_channel(app, federation_config, auth_plugin)
         stub = ExecStub(channel)  # pylint: disable=unused-variable # noqa: F841
 
         typer.secho(f"✋ Stopping run ID {run_id}...", fg=typer.colors.GREEN)
@@ -85,27 +83,6 @@ def stop(
         raise typer.Exit(code=1) from err
     finally:
         channel.close()
-
-
-def on_channel_state_change(channel_connectivity: str) -> None:
-    """Log channel connectivity."""
-    log(DEBUG, channel_connectivity)
-
-
-def _init_channel(app: Path, federation_config: dict[str, Any]) -> grpc.Channel:
-    """Initialize gRPC channel to the Exec API."""
-    insecure, root_certificates_bytes = validate_certificate_in_federation_config(
-        app, federation_config
-    )
-    channel = create_channel(
-        server_address=federation_config["address"],
-        insecure=insecure,
-        root_certificates=root_certificates_bytes,
-        max_message_length=GRPC_MAX_MESSAGE_LENGTH,
-        interceptors=None,
-    )
-    channel.subscribe(on_channel_state_change)
-    return channel
 
 
 def _stop_run(
